+++ src/pages/TodoPage.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import TodoList from 'components/TodoList';
-import TodoCreateForm from 'components/TodoCreateForm';
+import TodoList from 'components/todo/TodoList';
+import TodoCreateForm from 'components/todo/TodoCreateForm';
 import { getTodos } from 'api';
 import isLogin from 'utils/isLogin';
